MANTA-1936 Rip out syslog logging (#4)
MANTA-1936 Rip out syslog logging

Replace the use of syslog logging with SMF logging. This means that rather than have a single log file for all instances of the service that each instance logs to its own log file. Log rotation and upload is now managed via a new set of functions in the manta-scripts repository introduced as part of the changes for MANTA-4729 that ensure the SMF log files for each instance of the serviceare properly handled.